de-emphasize write timestamps

+++ libtiledbsoma/test/unit_soma_reader.cc
@@ -58,37 +58,50 @@ const std::string src_path = TILEDBSOMA_SOURCE_ROOT;
 
 namespace {
 
-uint64_t create_array(
-    const std::string& uri,
+std::tuple<std::string, uint64_t> create_array(
+    const std::string& uri_in,
     Context& ctx,
     int num_cells_per_fragment = 10,
     int num_fragments = 1,
     bool overlap = false,
-    bool allow_duplicates = false) {
-    // Delete array if it exists
-    auto vfs = VFS(ctx);
-    if (vfs.is_dir(uri)) {
-        vfs.remove_dir(uri);
+    bool allow_duplicates = false,
+    uint64_t timestamp = 1,
+    bool reuse_existing = false) {
+    std::string uri = fmt::format(
+        "{}-{}-{}-{}-{}",
+        uri_in,
+        num_cells_per_fragment,
+        num_fragments,
+        overlap,
+        allow_duplicates);
+    // Create array, if not reusing the existing array
+    if (!reuse_existing) {
+        auto vfs = VFS(ctx);
+        if (vfs.is_dir(uri)) {
+            vfs.remove_dir(uri);
+        }
+
+        // Create schema
+        ArraySchema schema(ctx, TILEDB_SPARSE);
+
+        auto dim = Dimension::create<int64_t>(
+            ctx, "d0", {0, std::numeric_limits<int64_t>::max()});
+
+        Domain domain(ctx);
+        domain.add_dimension(dim);
+        schema.set_domain(domain);
+
+        auto attr = Attribute::create<int>(ctx, "a0");
+        schema.add_attribute(attr);
+        schema.set_allows_dups(allow_duplicates);
+        schema.check();
+
+        // Create array
+        Array::create(uri, schema);
     }
 
-    // Create schema
-    ArraySchema schema(ctx, TILEDB_SPARSE);
-
-    auto dim = Dimension::create<int64_t>(
-        ctx, "d0", {0, std::numeric_limits<int64_t>::max()});
-
-    Domain domain(ctx);
-    domain.add_dimension(dim);
-    schema.set_domain(domain);
-
-    auto attr = Attribute::create<int>(ctx, "a0");
-    schema.add_attribute(attr);
-    schema.set_allows_dups(allow_duplicates);
-    schema.check();
-
-    // Create array and open for writing
-    Array::create(uri, schema);
-    Array array(ctx, uri, TILEDB_WRITE);
+    // Open array for writing
+    Array array(ctx, uri, TILEDB_WRITE, timestamp);
     if (LOG_DEBUG_ENABLED()) {
         array.schema().dump();
     }
@@ -127,7 +140,7 @@ uint64_t create_array(
         nnz = (num_fragments + 1) / 2 * num_cells_per_fragment;
     }
 
-    return nnz;
+    return {uri, nnz};
 }
 
 };  // namespace
@@ -143,16 +156,18 @@ TEST_CASE("SOMAReader: nnz") {
         num_fragments,
         overlap,
         allow_duplicates)) {
-        // Create array
-        std::string uri = "mem://unit-test-array";
         auto ctx = std::make_shared<Context>();
-        auto expected_nnz = create_array(
-            uri,
+
+        // Create array at timestamp 10
+        std::string base_uri = "mem://unit-test-array";
+        auto [uri, expected_nnz] = create_array(
+            base_uri,
             *ctx,
             num_cells_per_fragment,
             num_fragments,
             overlap,
-            allow_duplicates);
+            allow_duplicates,
+            10);
 
         // Get total cell num
         auto sr = SOMAReader::open(ctx, uri);
@@ -160,6 +175,119 @@ TEST_CASE("SOMAReader: nnz") {
         uint64_t nnz;
         if (num_fragments > 1 && overlap && allow_duplicates) {
             REQUIRE_THROWS(nnz = sr->nnz());
+            LOG_DEBUG("Caught expected exception for nnz with duplicates");
+        } else {
+            nnz = sr->nnz();
+            REQUIRE(nnz == expected_nnz);
+        }
+    }
+}
+
+TEST_CASE("SOMAReader: nnz with timestamp") {
+    auto num_fragments = GENERATE(1, 10);
+    auto overlap = GENERATE(false, true);
+    auto allow_duplicates = GENERATE(false, true);
+    int num_cells_per_fragment = 128;
+
+    SECTION(fmt::format(
+        " - fragments={}, overlap={}, allow_duplicates={}",
+        num_fragments,
+        overlap,
+        allow_duplicates)) {
+        auto ctx = std::make_shared<Context>();
+
+        // Create array at timestamp 10
+        std::string base_uri = "mem://unit-test-array";
+        const auto& [uri, expected_nnz] = create_array(
+            base_uri,
+            *ctx,
+            num_cells_per_fragment,
+            num_fragments,
+            overlap,
+            allow_duplicates,
+            10);
+
+        // Write more data to the array at timestamp 20, which will be
+        // not be included in the nnz call with a timestamp
+        create_array(
+            base_uri,
+            *ctx,
+            num_cells_per_fragment,
+            num_fragments,
+            overlap,
+            allow_duplicates,
+            20,
+            true);
+
+        // Get total cell num at timestamp (0, 15)
+        std::pair<uint64_t, uint64_t> timestamp{0, 15};
+        auto sr = SOMAReader::open(
+            ctx, uri, "nnz", {}, "auto", "auto", timestamp);
+
+        uint64_t nnz;
+        if (num_fragments > 1 && overlap && allow_duplicates) {
+            REQUIRE_THROWS(nnz = sr->nnz());
+            LOG_DEBUG("Caught expected exception for nnz with duplicates");
+        } else {
+            nnz = sr->nnz();
+            REQUIRE(nnz == expected_nnz);
+        }
+    }
+}
+
+TEST_CASE("SOMAReader: nnz with timestamp and consolidation") {
+    auto num_fragments = GENERATE(1, 10);
+    auto overlap = GENERATE(false, true);
+    auto allow_duplicates = GENERATE(false, true);
+    auto vacuum = GENERATE(false, true);
+    int num_cells_per_fragment = 128;
+
+    SECTION(fmt::format(
+        " - fragments={}, overlap={}, allow_duplicates={}, vacuum={}",
+        num_fragments,
+        overlap,
+        allow_duplicates,
+        vacuum)) {
+        auto ctx = std::make_shared<Context>();
+
+        // Create array at timestamp 10
+        std::string base_uri = "mem://unit-test-array";
+        const auto& [uri, expected_nnz] = create_array(
+            base_uri,
+            *ctx,
+            num_cells_per_fragment,
+            num_fragments,
+            overlap,
+            allow_duplicates,
+            10);
+
+        // Write more data to the array at timestamp 20, which will be
+        // duplicates of the data written at timestamp 10
+        create_array(
+            base_uri,
+            *ctx,
+            num_cells_per_fragment,
+            num_fragments,
+            overlap,
+            allow_duplicates,
+            20,
+            true);
+
+        // Consolidate and optionally vacuum
+        Array::consolidate(*ctx, uri);
+        if (vacuum) {
+            Array::vacuum(*ctx, uri);
+        }
+
+        // Get total cell num
+        std::pair<uint64_t, uint64_t> timestamp{1, 15};
+        auto sr = SOMAReader::open(
+            ctx, uri, "nnz", {}, "auto", "auto", timestamp);
+
+        uint64_t nnz;
+        if (allow_duplicates) {
+            REQUIRE_THROWS(nnz = sr->nnz());
+            LOG_DEBUG("Caught expected exception for nnz with duplicates");
         } else {
             nnz = sr->nnz();
             REQUIRE(nnz == expected_nnz);
